+++ test/Generics/unify_nested_types_5.swift
@@ -1,4 +1,4 @@
-// RUN: %target-typecheck-verify-swift -enable-requirement-machine
+// RUN: %target-typecheck-verify-swift -requirement-machine=verify
 
 protocol P {
   associatedtype A : P
